tank/requirements.txt


# tankdemo/app.py: 8
Pillow == 7.0.0

# tankdemo/app.py: 7
altair == 4.1.0

# tankdemo/app.py: 3,4
matplotlib == 3.3.3

# tankdemo/app.py: 2
pandas == 1.2.4

# tankdemo/app.py: 6
seaborn == 0.11.1

# tankdemo/app.py: 1
streamlit == 0.82.0
